clickAttackDamage, clickAttackSpeed, canClickAttack, clickTimer and clickTimerMax added to player object. (13)
updateEnemyContainer() added, will display a CSS class based on if the player can click attack or not (this is an indicator to the player) (228)

Added enemyImageContainer const. (251)

Added functionClickAttackEnemy, this checks if the player is able to click attack, if they are, the enemy is damaged and a message is logged to the console. Once clicked, the player object's canClickAttack is set to false and we call the updateEnemyContainer() function, we then setTimeout for player object's clickAttackSpeed * 1000 and then set attack.CanClickAttack to true. Once it gets to true, we call updateEnemyContainer. (374)

Added event listener to enemyImage (394)