Added changeFloor function, if player is in combat, function returns. If direction give is ascend, player.floor ++, direct is descend, floor -- unless current floor is 0. (94 - 106)
Added updateDOM() function with TODO for later, will just throw all DOM updates into there. (255-256)

Added ascendFloorButton and descendFloorButton selectors. (282-285)

Attached event listeners to ascendFloorButton and descendFloorButton, if the player clicks ascend, changeFloor is called and passed the value "ascend", if descend is clicked, changeFloor is called and the passed value is "descend". (395-406)